pyproject.toml
[project]
name = "sportpools-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawlee[all]>=0.6.9",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
]
